Merged in reverse-loop-pattern-method-lookup (pull request #9)
Reverse loop pattern method lookup

* Add test case fail message

* Filter test case match on root package

* Improve test case identification

This version walks the stack trace in reverse and uses a regex to match
the test case method.

A new argument to the agent has been added. MethodmapTransformer uses a
package name different from the package loader:

- the package loader loads classes in the target project that should be
  instrumented whereas the transformer uses the package name to look for
  test cases.
- As in the case of dynamic agent tests and static agent tests it may be
  useful to separate them.

But it may be redundant. In any case we can fallback to use the same
package name for both if only one is specified.

* Try a new method of throwing IOException

* Use a mock Path to force IOException

* Refactor

Approved-by: Isak Kristola